requirements.txt
torch>=1.5.0
sphinx>=2.1.2
networkx>=2.2
setuptools>=40.8.0
matplotlib>=3.0.3
numpy>=1.16.2
seaborn>=0.9.0
scikit_learn>=0.20.3
hyperopt>=0.2.1
pathlib>=1.0.1
numpydoc>=0.9.1
sphinx-gallery>=0.3.1
sphinx-rtd-theme>=0.4.3
pytest>=3.6
pyyaml>=5.3.1
click